+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased](https://github.com/botopolis/slack/compare/v0.4.0...master)
+
+### Added
+
+- Reactions: `Adapter.React(bot.Message)`. React to messages with emoji.
+
 ## [0.4.0](https://github.com/botopolis/slack/compare/v0.3.1...v0.4.0) - 2018-06-24
 
 ### Added